Big rework of the change submission and model-update triggering.
Now we don't wait for files to be saved anymore, but changes are
notified to the central model as soon as they are committed. However the
central model might not be updated immediately, which doesn't prevent
the local typechecking to occur in the meantime. Change submission and
model-update triggering are managed asynchronously for obvious
performance reasons.