Make the typecheckerAnnotator an `ExternalAnnotator`
so that the possibly long typechecking phase is *not* done inside a
global `Read-access` section. This reduces the risk of live-locks.
This also allows marking this annotator as
DumbAware, so that the alternate resolution mechanism will be used for
any local typechecking that would occur when indexes are not available.