Adapt to changes in the `CeylonFile` API:
Also pass the `phasedUnit` in the quickFixes function directly, since
the `LocalAnalysisResult` might change at any moment if a new local
typechecking starts in the meantime and, on the other hand, using
`lastPhasedUnit`might use an obsolete phased unit.

In the context of #427 (Provide a non-blocking way to wait for an
up-to-date analysis result/phased unit associated to a `CeylonFile`).