Fi a performance regression.
The change from `CeylonFile.classes` to use the `CeylonSourceNavigator` had the bad side-effect
to involve a wait for an up-to-date (lazily typechecked) external phased unit as soon as the navigation class
was requested, which means way more often than really required. This slowed down the display of the
`Navigate > Class` quite much, even blocking it in case of model update.

In the context of #442 (CeylonFile.waitForUpToDatePhasedUnit() is super slow)